+++ Tests/Unit/Templating/Fixtures/TestingConfigurationCheck.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace OliverKlee\Oelib\Tests\Unit\Templating\Fixtures;
+
+/**
+ * Testing configuration check.
+ *
+ * @author Oliver Klee <typo3-coding@oliverklee.de>
+ */
+class TestingConfigurationCheck extends \Tx_Oelib_ConfigCheck
+{
+}